# documentation/api.yaml
openapi: 3.0.3
info:
  title: Arasvara News Platform API
  description: API untuk platform berita Arasvara - CMS, article management, analytics, dan advertising
  version: 1.0.0
  contact:
    name: Arasvara Tech Team
    email: tech@arasvara.com
servers:
  - url: https://api.arasvara.com
    description: Production server
  - url: https://staging-api.arasvara.com
    description: Staging server
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...
paths:
...
  /api/analytics/view-article:
    post:
      tags:
        - Analytics
      summary: Track article view
      description: Public endpoint untuk tracking view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - articleId
              properties:
                articleId:
                  type: string
                sessionId:
                  type: string
                referrer:
                  type: string
      responses:
        '200':
          description: View tracked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # Search APIs
...